+++ src/modules/esm.mjs
@@ -1,0 +1,41 @@
+import path from 'path'
+import {release, version} from 'os'
+import {fileURLToPath} from 'url'
+import { createServer as createServerHttp } from "http"
+import './files/c.js'
+import {dirname, join} from 'path'
+
+const __fileName = fileURLToPath(import.meta.url)
+const __dirName = dirname(__fileName)
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = await import("./files/a.json", {with:{type: 'json'}});
+} else {
+    unknownObject =  await import("./files/b.json", {with:{type: 'json'}});
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__fileName}`);
+console.log(`Path to current directory is ${__dirName}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {unknownObject, myServer}
